# docs/openapi.yml
openapi: 3.1.0
info:
  title: fast-auth
  description: A simple, extensible authentication library for Axum with JWT and refresh tokens
  license:
    name: Apache-2.0
    identifier: Apache-2.0
  version: 0.1.0
paths:
  /auth/api-keys:
    get:
      tags:
      - auth
      summary: List API keys owned by the current user.
      operationId: api_keys_list
      parameters:
      - name: limit
        in: query
        description: Maximum number of items to return.
        required: true
        schema:
          type: integer
          format: int32
          default: 20
          maximum: 100
          minimum: 1
      - name: offset
        in: query
        description: Number of items to skip from the result set.
        required: true
        schema:
          type: integer
          format: int32
          default: 0
          minimum: 0
      - name: sort_by
        in: query
        description: Column used for sorting.
        required: true
        schema:
          $ref: '#/components/schemas/ApiKeyListSortBy'
      - name: sort_order
        in: query
        description: Sort direction.
        required: true
        schema:
          $ref: '#/components/schemas/ListSortOrder'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListPageResult_ApiKeySummary'
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
        '500':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
      security:
      - sessionCookie: []
      - bearerApiKey: []
    post:
      tags:
      - auth
      summary: Create one API key for the current organization.
      operationId: api_key_create
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApiKeyCreateRequest'
        required: true
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyCreateResponse'
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
        '500':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
      security:
      - sessionCookie: []
      - bearerApiKey: []
  /auth/api-keys/{api_key_id}:
    delete:
      tags:
      - auth
      summary: Delete one API key owned by the current user.
      operationId: api_key_delete
      parameters:
      - name: api_key_id
        in: path
        description: API key id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeySummary'
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
        '404':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
        '500':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
      security:
      - sessionCookie: []
      - bearerApiKey: []
  /auth/email/confirm:
    get:
      tags:
      - auth
      summary: Confirm email from a browser verification link (`GET /auth/email/confirm?token=...`).
      operationId: email_confirm_get
      parameters:
      - name: token
        in: query
        description: Verification token from email link.
        required: true
        schema:
          type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmailConfirmResponse'
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
        '500':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
  /auth/email/confirm/send:
    post:
      tags:
      - auth
      summary: Send a confirmation email to the user.
      description: |-
        Creates a verification token and sends an email with the confirmation link.
        Always returns success to prevent email enumeration attacks.
      operationId: email_confirm_send
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailConfirmSendRequest'
        required: true
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmailConfirmSendResponse'
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
        '500':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
...
components:
  schemas:
    ApiKeyCreateRequest:
      type: object
      description: Request body for API key creation.
      required:
      - name
      properties:
        name:
          type: string
          description: User-defined API key name.
    ApiKeyCreateResponse:
      type: object
      description: API key creation response with one-time plaintext key.
      required:
      - id
      - name
      - organization_id
      - key
      - created_by_user_id
      - key_prefix
      - created_at
      properties:
        created_at:
          type: string
          format: date-time
          description: Creation timestamp.
        created_by_user_id:
          type: string
          format: uuid
          description: User id that created the key.
        id:
          type: string
          format: uuid
          description: API key identifier.
        key:
          type: string
          description: Plaintext key returned once.
        key_prefix:
          type: string
          description: Stable visible key prefix.
        last_used_at:
          type:
          - string
          - 'null'
          format: date-time
          description: Last successful use timestamp.
        name:
          type: string
          description: User-defined display name.
        organization_id:
          type: string
          format: uuid
          description: Owning organization id.
    ApiKeyListSortBy:
      type: string
      description: Sortable columns for API-key listing.
      enum:
      - created_at
      - name
      - last_used_at
    ApiKeySummary:
      type: object
      description: API key summary returned by list and delete endpoints.
      required:
      - id
      - name
      - organization_id
      - key_prefix
      - created_by_user_id
      - created_at
      properties:
        created_at:
          type: string
          format: date-time
          description: Creation timestamp.
        created_by_user_id:
          type: string
          format: uuid
          description: User id that created the key.
        id:
          type: string
          format: uuid
          description: API key identifier.
        key_prefix:
          type: string
          description: Stable visible key prefix.
        last_used_at:
          type:
          - string
          - 'null'
          format: date-time
          description: Last successful use timestamp.
        name:
          type: string
          description: User-defined display name.
        organization_id:
          type: string
          format: uuid
          description: Owning organization id.
    AuthErrorResponse:
      type: object
      description: Standard error response body returned by auth endpoints.
      required:
      - error
      properties:
        error:
          type: string
          description: Human-readable error message.
...
    EmailConfirmResponse:
      type: object
      description: Response for confirming email.
      required:
      - message
      properties:
        message:
          type: string
          description: Success message.
    EmailConfirmSendRequest:
      type: object
      description: Request body for sending confirmation email.
      required:
      - email
      properties:
        email:
          type: string
          description: User's email address.
    EmailConfirmSendResponse:
      type: object
      description: Response for sending confirmation email.
      required:
      - message
      properties:
        message:
          type: string
          description: Success message.
    ListPageResult_ApiKeySummary:
      type: object
      description: Shared page payload metadata plus returned items.
      required:
      - items
      - total
      - limit
      - offset
      properties:
        items:
          type: array
          items:
            type: object
            description: API key summary returned by list and delete endpoints.
            required:
            - id
            - name
            - organization_id
            - key_prefix
            - created_by_user_id
            - created_at
            properties:
              created_at:
                type: string
                format: date-time
                description: Creation timestamp.
              created_by_user_id:
                type: string
                format: uuid
                description: User id that created the key.
              id:
                type: string
                format: uuid
                description: API key identifier.
              key_prefix:
                type: string
                description: Stable visible key prefix.
              last_used_at:
                type:
                - string
                - 'null'
                format: date-time
                description: Last successful use timestamp.
              name:
                type: string
                description: User-defined display name.
              organization_id:
                type: string
                format: uuid
                description: Owning organization id.
          description: Returned page items.
        limit:
          type: integer
          format: int32
          description: Applied page size.
          maximum: 100
          minimum: 1
        offset:
          type: integer
          format: int32
          description: Applied page offset.
          minimum: 0
        total:
          type: integer
          format: int64
          description: Total number of visible rows.
    ListSortOrder:
      type: string
      description: Shared sort direction for list endpoints.
      enum:
      - asc
      - desc
...
  securitySchemes:
    bearerApiKey:
      type: apiKey
      in: header
      name: Authorization
    sessionCookie:
      type: apiKey
      in: cookie
      name: access_token
...